ECS linking, working on GC
Tests in EntityTest show the linking between entities and components is
generally working as wanted with a minimal public API.

Current issue is that calling `Destroy()` on the entity doesn't seem to
actually destroy it. Not clear where reference to entity is. Debugging
hasn't shown it up.

Try the IntelliJ memory unit framework maybe